moving all the logging middleware stuff in its own file
