+++ src/internal/middleware/middleware.go
@@ -1,25 +1,13 @@
 package middleware
 
-import (
-  "time"
-  "net/http"
 
-  "github.com/BeringLogic/flimsy/internal/logger"
+import (
+  "net/http"
 )
 
 
-type wrappedWriter struct {
-  http.ResponseWriter
-  statusCode int
-}
-
 type Middleware func(http.Handler) http.Handler
 
-
-func (w *wrappedWriter) WriteHeader(statusCode int) {
-  w.ResponseWriter.WriteHeader(statusCode)
-  w.statusCode = statusCode
-}
 
 func CreateStack(m ...Middleware) Middleware {
   return func(next http.Handler) http.Handler {
@@ -29,18 +17,3 @@ func CreateStack(m ...Middleware) Middleware {
     return next
   }
 }
-
-func Logging(log logger.FlimsyLogger, next http.Handler) http.Handler {
-  return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-    start := time.Now()
-
-    wrapped := &wrappedWriter{
-      ResponseWriter: w,
-      statusCode: http.StatusOK,
-    }
-
-    next.ServeHTTP(wrapped, r)
-    log.Printf("| %d | %s | %s | %s", wrapped.statusCode, r.Method, r.URL.Path, time.Since(start))
-  })
-}
-
